pyproject.toml
[tool.poetry]
name = "glassdoorinterviewexpert"
version = "0.1.0"
description = "Scrapes interview questions from glassdoor for a specified company and uses OpenAIs API to summarize them."
authors = ["Franzi <92818644+franziskavonalbedyll@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
undetected-chromedriver = "^3.5.4"
langchain = "^0.0.340"
python-dotenv = "^1.0.0"
openai = "^1.3.5"
chromadb = "^0.4.18"
tiktoken = "^0.5.1"
mdutils = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
